Projects/create_your_own_chatbot_streamlit/pyproject.toml
[project]
name = "create-your-own-chatbot-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.4.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
